+++ posthog/models/filters/mixins/common.py
@@ -38,8 +38,8 @@ from posthog.constants import (
 )
 from posthog.models.entity import Entity, ExclusionEntity
 from posthog.models.filters.mixins.base import BaseParamMixin, BreakdownType, IntervalType
-from posthog.models.filters.mixins.groups import validate_group_type_index
 from posthog.models.filters.mixins.utils import cached_property, include_dict, process_bool
+from posthog.models.filters.utils import validate_group_type_index
 from posthog.utils import relative_date_parse
 
 ALLOWED_FORMULA_CHARACTERS = r"([a-zA-Z \-\*\^0-9\+\/\(\)]+)"
