+++ posthog/models/filters/utils.py
@@ -1,8 +1,10 @@
-from typing import Optional
+from typing import Any, Optional
 
+from rest_framework.exceptions import ValidationError
 from rest_framework.request import Request
 
 from posthog.constants import (
+    GROUP_TYPES_LIMIT,
     INSIGHT_FUNNELS,
     INSIGHT_PATHS,
     INSIGHT_RETENTION,
@@ -46,3 +48,22 @@ def get_filter(team, data: dict = {}, request: Optional[Request] = None):
             data={**data, **(request.data if request else {}), "insight": INSIGHT_FUNNELS}, request=request, team=team
         )
     return Filter(data=data, request=request, team=team)
+
+
+def validate_group_type_index(param_name: str, value: Any, required=False) -> Optional[int]:
+    error = ValidationError(
+        f"{param_name} is required to be greater than 0 and less than {GROUP_TYPES_LIMIT}", code="invalid"
+    )
+
+    if required and value is None:
+        raise error
+
+    if value is not None:
+        try:
+            value = int(value)
+        except:
+            raise error
+        if not (0 <= value < GROUP_TYPES_LIMIT):
+            raise error
+
+    return value
